search: move search code
